Refactor experimental tensorboard Hparams:
