+++ maggy/tensorboard.py
@@ -1,9 +1,8 @@
-import tensorflow as tf
 
 from maggy.core import config
 
-if config.mode is config.HOPSWORKS:
-    from tensorboard.plugins.hparams import api as hp
+if config.mode is config.HOPSWORKS and config.tf_version >= 2:
+    from maggy.tb_experimental import write_hparams_proto
 
 tensorboard_dir = None
 
@@ -15,47 +14,3 @@ def _register(trial_dir):
 def logdir():
     global tensorboard_dir
     return tensorboard_dir
-
-def create_hparams_config(sp):
-    hparams = []
-
-    for k, v in sp.names().items():
-        if v == 'DOUBLE':
-            hparams.append(hp.HParam(k, hp.RealInterval(float(sp.get(k)[0]), float(sp.get(k)[1]))))
-        elif v == 'INTEGER':
-            hparams.append(hp.HParam(k, hp.IntInterval(sp.get(k)[0], sp.get(k)[1])))
-        elif v == 'DISCRETE':
-            hparams.append(hp.HParam(k, hp.Discrete(sp.get(k))))
-        elif v == 'CATEGORICAL':
-            hparams.append(hp.HParam(k, hp.Discrete(sp.get(k))))
-
-    return hparams
-
-def write_hparams_proto(logdir, searchspace):
-
-    HPARAMS = create_hparams_config(searchspace)
-    METRICS = [
-        hp.Metric(
-            "epoch_accuracy",
-            group="validation",
-            display_name="accuracy (val.)",
-        ),
-        hp.Metric(
-            "epoch_loss",
-            group="validation",
-            display_name="loss (val.)",
-        ),
-        hp.Metric(
-            "batch_accuracy",
-            group="train",
-            display_name="accuracy (train)",
-        ),
-        hp.Metric(
-            "batch_loss",
-            group="train",
-            display_name="loss (train)",
-        ),
-    ]
-
-    with tf.summary.create_file_writer(logdir).as_default():
-        hp.hparams_config(hparams=HPARAMS, metrics=METRICS)
